Zmiana sterowania: teraz oprócz samych żądań jazdy na danej osi, symulacja obsługuje żądanie wykorzystania tylko części maksymalnych możliwości samochodu odnośnie skrętu/prędkości.
Dodatkowo zmeinione SingleTouchUI na JoystickSingleTouchUI, dżojstik przeniesiony w prawy dolny róg ekranu (wygodny dostęp kciukiem przy trzymaniu telefonu jedną ręką, przynajmniej dla praworęcznych), zmieniona grafika i dodana współpraca z nową symulacją.

+++ src/com/games/racertc/ui/UIManager.java
@@ -24,10 +24,10 @@ public abstract class UIManager implements OnTouchListener, GameStateChangeListe
 	protected MessageFactory messageFactory;
 	
 	/** Szerokosc ekranu. */
-	protected int width = 1024;
+	protected float width = 1024;
 	
 	/** Wysokosc ekranu. */
-	protected int height = 768;
+	protected float height = 768;
 	
 	/**
 	 * Konstruktor klasy buforujacy obiekty MessageQueue i MessageFactory. Powinien
@@ -47,8 +47,8 @@ public abstract class UIManager implements OnTouchListener, GameStateChangeListe
 	 * @param height Nowa wysokosc ekranu.
 	 */
 	public void setResolution( int width, int height ) {
-		this.width = width;
-		this.height = height;
+		this.width = (float) width;
+		this.height = (float) height;
 	}
 	
 /*-----------------------------------------*/
